Remember the scrollbar position when clicking a QM toolbar link.
If the scroll position is above the Query Monitor panel when a toolbar
link is clicked, it will be saved so that clicking the same link will
return to the original scroll position. Clicking the overview link while
viewing the QM panel also returns to the previous scroll position.

This makes it easier to quickly view a notice and return to the previous
position without having to scroll back up a couple thousand pixels if
the notice can be ignored.