Add labels to filters
Filter fields become more accessible when they have a label that is
correctly marked up.

This also stops QM from being the source of errors when checking a page
with a tool like the WAVE Evaluation Tool (browser extension).

Originally added with a new `build_filter_label()` method, but this then
created a redundant `<label>` when no filter field was present.