Adjust colour contrasts
#767676 is the minimum grey that can be used on the white background, so
"info" and "false" statements need to use this. This does unfortunately
leave only a very minor subtle difference in colour between main and
secondary text, but better that than none of the secondary text being
readable by some users.

The "true" styles relied on changing the text colour only. This is now
shown with a green background and white text, which should be more
noticeable even though it is still relying on just colour.