Table a11y: Assets
This was previously a single table with "totally-legit-spacer" to visually break up the scripts and styles.
This commit changes it to two tables, legitimately spaced as per all the other tables. This allows a caption per table, and correct column scoping.

Two new column headings added, and previous content for the new headings is moved to the caption.