Table a11y: Admin
Additional table headings are for screen-readers only.

This previously had a nested table, but this commit does away with the nested table, and just counts the public object properties of `get_current_screen()` to know how many rows the table heading should span.

`scope` attributes have NOT been added in this commit, as it believed that this is sufficiently complex that `id` and `headers` attributes would be needed to correctly identify data with headings.