WP 5.2 may result in empty `$filtered_trace`
In WP 5.2 a plugin that doesn't meet the minimum PHP requirements will throw a WP_Die error. When tracing this in QM there there is only a single item in $filtered_trace. After the `array_pop( $filtered_trace );` the value of $filtered_trace is an empty array.

This PR accounts for that edge case.