Remove support for multiple instances of `wpdb`.
This feature is little used, I've not used it myself in the last 8 years. As nifty as it is, it contributes significantly to the maintenance burden of QM itself and slows down my ability to work on the front end rendering tasks. Maybe it'll reappear in the future, but it's unlikely.